NAVIGATION CON UNA SELED CLASS PARA LOS PARAMETROS, Y CREO PAQUETES PARA CADA PANTALLA
