+++ app/src/main/java/com/jfg/gitpruebas/presentation/bombitas/BombitasScreen.kt
@@ -1,6 +1,5 @@
-package com.jfg.gitpruebas.presentation
+package com.jfg.gitpruebas.presentation.bombitas
 
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -10,6 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
+import com.jfg.gitpruebas.presentation.navigation.Routes
 
 @Composable
 fun Bombitas(name: String, modifier: Modifier = Modifier,controller: NavHostController) {
@@ -24,7 +24,7 @@ fun Bombitas(name: String, modifier: Modifier = Modifier,controller: NavHostCont
                 modifier = modifier
         )
         Text(text = "BOmbitas")
-        Button(onClick = { controller.navigate("screen1") }) {
+        Button(onClick = { controller.navigate(Routes.Screen1.route) }) {
             Text(text = "Screen1")
         }
     }
