+++ app/src/main/java/com/jfg/gitpruebas/presentation/screen1/Screen1.kt
@@ -1,4 +1,4 @@
-package com.jfg.gitpruebas.presentation
+package com.jfg.gitpruebas.presentation.screen1
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -13,19 +13,19 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.navigation.NavHostController
+import com.jfg.gitpruebas.presentation.navigation.Routes
 
 @Composable
-fun Screen2(controller: NavHostController) {
-    Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color.Blue),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
+fun Screen1(controller: NavHostController) {
+    Column(modifier = Modifier
+        .fillMaxSize()
+        .background(Color.Blue),
+           horizontalAlignment = Alignment.CenterHorizontally,
+           verticalArrangement = Arrangement.Center
     ) {
-        Text(text = "Screen2")
-        Button(onClick = { controller.navigate("bombitas") }) {
-            Text(text = "Bombitas")
+        Text(text = "Screen1")
+        Button(onClick = { controller.navigate(Routes.Screen2.route) }) {
+            Text(text = "Screen2")
         }
     }
 }
